move paginationtext to components
